Add files via upload
The banner for Task 1, this banner is eye catching, with all the required features. my home town has no motto of its own, so i decided to use the motto of the primary school located there instead, as to me this seems the most appropriate. The colour used will be used as a potential inspiration for Task 2