Add files via upload
For Task 5, i have added descriptions for each page, as well as images for the areas to visit. The pictures are references in each section, and the layout is easy to follow.  There is a link to the website where i got the information for the history in the webpage.